feat: create abstract job manager and local job manager
