+++ EasyCLI/Commands/ListCommand.cs
@@ -23,7 +23,7 @@ public class ListCommand : Command
     public override void Run(IEnumerable<string> args)
     {
         var argsList = args.ToList();
-        var jm = new JobManager();
+        var jm = new LocalJobManager();
         var fetchSuccess = jm.FetchJobs();
 
         if (!fetchSuccess)
