+++ EasyCLI/Commands/RunCommand.cs
@@ -57,7 +57,7 @@ public class RunCommand : Command, IJobStatusSubscriber
             return;
         }
 
-        var jm = new JobManager();
+        var jm = new LocalJobManager();
         var fetchSuccess = jm.FetchJobs();
 
         if (!fetchSuccess)
